+++ tests-corpus/valid-bare-schema/alias-doccomment/out.gen.js
@@ -1,0 +1,17 @@
+import * as bare from "@bare-ts/lib"
+
+export function readU8Alias(bc) {
+    return bare.readU8(bc)
+}
+
+export function writeU8Alias(bc, x) {
+    bare.writeU8(bc, x)
+}
+
+export function readU8Alias2(bc) {
+    return bare.readU8(bc)
+}
+
+export function writeU8Alias2(bc, x) {
+    bare.writeU8(bc, x)
+}