+++ tests-corpus/valid-bare-schema/enum-doccomment/out.gen.ts
@@ -1,0 +1,48 @@
+import * as bare from "@bare-ts/lib"
+
+/**
+ * An enum to model Genders
+ */
+export enum Gender {
+    /**
+     * Be inclusive :)
+     */
+    FLUID = "FLUID",
+    MALE = "MALE",
+    /**
+     * One is not born, but becomes a woman
+     *                  -- Simone de Beauvoir
+     */
+    FEMALE = "FEMALE",
+}
+
+export function readGender(bc: bare.ByteCursor): Gender {
+    const offset = bc.offset
+    const tag = bare.readU8(bc)
+    switch (tag) {
+        case 0:
+            return Gender.FLUID
+        case 1:
+            return Gender.MALE
+        case 2:
+            return Gender.FEMALE
+        default: {
+            bc.offset = offset
+            throw new bare.BareError(offset, "invalid tag")
+        }
+    }
+}
+
+export function writeGender(bc: bare.ByteCursor, x: Gender): void {
+    switch (x) {
+        case Gender.FLUID:
+            bare.writeU8(bc, 0)
+            break
+        case Gender.MALE:
+            bare.writeU8(bc, 1)
+            break
+        case Gender.FEMALE:
+            bare.writeU8(bc, 2)
+            break
+    }
+}