+++ tests-corpus/valid-bare-schema/struct-doccomment/out.gen.d.ts
@@ -1,0 +1,21 @@
+import * as bare from "@bare-ts/lib"
+
+export type u8 = number
+
+/**
+ * A struct to model persons
+ */
+export interface Person {
+    /**
+     * person's name
+     */
+    readonly name: string
+    /**
+     * person's age
+     */
+    readonly age: u8
+}
+
+export function readPerson(bc: bare.ByteCursor): Person
+
+export function writePerson(bc: bare.ByteCursor, x: Person): void