+++ tests-corpus/valid-bare-schema/struct-doccomment/out.gen.ts
@@ -1,0 +1,29 @@
+import * as bare from "@bare-ts/lib"
+
+export type u8 = number
+
+/**
+ * A struct to model persons
+ */
+export interface Person {
+    /**
+     * person's name
+     */
+    readonly name: string
+    /**
+     * person's age
+     */
+    readonly age: u8
+}
+
+export function readPerson(bc: bare.ByteCursor): Person {
+    return {
+        name: bare.readString(bc),
+        age: bare.readU8(bc),
+    }
+}
+
+export function writePerson(bc: bare.ByteCursor, x: Person): void {
+    bare.writeString(bc, x.name)
+    bare.writeU8(bc, x.age)
+}